Refactored MinQuantity
