+++ src/main/java/seedu/fridgefriend/command/RunningLowCommand.java
@@ -3,6 +3,7 @@ package seedu.fridgefriend.command;
 //@@author kwokyto
 import seedu.fridgefriend.exception.InvalidQuantityException;
 import seedu.fridgefriend.food.FoodCategory;
+import seedu.fridgefriend.food.MinimumQuantity;
 import seedu.fridgefriend.utilities.Ui;
 
 public class RunningLowCommand extends Command {
@@ -25,7 +26,7 @@ public class RunningLowCommand extends Command {
     @Override
     public void execute() throws InvalidQuantityException {
         for (FoodCategory foodCategory : FoodCategory.values()) {
-            if (FoodCategory.getMinimumQuantity(foodCategory) == TURN_OFF) {
+            if (MinimumQuantity.getMinimumQuantity(foodCategory) == TURN_OFF) {
                 numberOfCategoryTurnOff++;
                 continue;
             }
@@ -45,7 +46,7 @@ public class RunningLowCommand extends Command {
             isStockedUp = false;
             int totalQuantity = fridge.getTotalQuantity(foodCategory);
             String entry = "\n" + index + ". " + foodCategory.toString() + " quantity: " + totalQuantity
-                    + " out of " + FoodCategory.getMinimumQuantity(foodCategory);
+                    + " out of " + MinimumQuantity.getMinimumQuantity(foodCategory);
             message += entry;
             index += 1;
         }
